[cmake] Use whole-archive instead of force_load with Clang
This fixes linker errors when building with Clang on Linux. Only
AppleClang understand force_load, Clang on Linux needs to use
whole-archive like gcc.

This is discussed in cmake documentation here
https://cmake.org/cmake/help/latest/variable/CMAKE_LANG_LINK_LIBRARY_USING_FEATURE.html#loading-a-whole-static-library
and one can also look at how the WHOLEARCHIVE feature is handled by
CMake here
https://github.com/search?q=whole_archive+repo%3AKitware%2FCMake+path%3A%2FModules

It is possible that Clang on Windows should use /WHOLEARCHIVE and not
--whole-archive but I can't test that easily.